Got very basic relief mapping working! Basically just updates to the shaders, loading an image and minor fixes.
Forgot to mention in the previous commit message, SFML does not seem to expose the necessary code to use the OpenGL extensions for shaders (specifically, GraphicsContext to initialise GLEW). I got access to this by adding the SFML source directory as an include directory, which is not ideal but it works.